requirements.txt
spotipy
python-dotenv
Pillow
requests